Add radius search to drivers/reports
Radius is not required for search but if it is provided, it is restricted
to a minimum of 1 mile. User is informed if they try to submit values
less than minimum. Search gets the coordinates from the provided address
then queries the db using those coords and provided radius and returns a
ransacked filtering of the results.

tg-01