Fix issues with database and searching and add info to markers
We now store current_state and current_city fields which allows us
to perform ransack searches on those fields. Reporting is updated to
use these new fields in conjunction with radius searches

Bonus: Fix / improve state report. The state report is now triggered
via right click. If a report is attempted for an unknown region,
e.g. middle of the ocean, for now we are just displaying an
'Unknown Region' report with no drivers.

This patch also hides a strange map marker error that was occuring
every so often on page load. The markers appeared to work as expected
despite this error so for now we are just ignoring this

TG-12
TG-14