Add live feed update
The feed is updated live using websockets. To do this required
upgrading the entire project to Rails 5 and making use of ActionCables.
There is a significant amount of refactoring to make this happen. All of the
logic for rendering markers and table rows based on websocket messages
can be found in assets/javascripts/channels/drivers.js